requirements/local.txt
Django==2.2.3
psycopg2==2.7.4
pytz==2019.1
sqlparse==0.3.0
djangorestframework==3.9.4
django-environ==0.4.5
Pillow==6.2.1
django-filter==2.2.0